pyproject.toml
[project]
name = "eshop"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Bill", email = "billxs@hotmail.com" }
]
dependencies = [
    "flask>=3.1.0",
    "python-dotenv>=1.1.0",
    "waitress>=3.0.2",
    "flask-sqlalchemy>=3.1.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.0.7",
    "flask-bcrypt>=1.0.1",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "numpy>=2.2.6",
    "redis>=6.2.0",
    "flask-wtf>=1.2.2",
    "faker>=37.4.0",
    "matplotlib>=3.10.3",
    "seaborn>=0.13.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/eshop"]


[tool.rye.scripts]
devserver = "flask --app src/eshop run"
dev = "waitress-serve --port=5000 --threads=4 eshop:app"
prod = "waitress-serve --port=80 --threads=4 eshop:app"
